+++ src/main/java/com/nemonotfound/nemoscarpentry/block/ModBlocks.java
@@ -132,6 +132,28 @@ public class ModBlocks {
             new LukasChairBlock(AbstractBlock.Settings.copy(SPRUCE_PLANKS).nonOpaque()));
     public static final Block WARPED_CHAIR_LUKAS = registerBlock("warped_chair_lukas",
             new LukasChairBlock(AbstractBlock.Settings.copy(WARPED_PLANKS).nonOpaque()));
+    public static final Block ACACIA_CHAIR_LUKAS_NATURAL = registerBlock("acacia_chair_lukas_natural",
+            new LukasChairBlock(AbstractBlock.Settings.copy(ACACIA_PLANKS).nonOpaque()));
+    public static final Block BAMBOO_CHAIR_LUKAS_NATURAL = registerBlock("bamboo_chair_lukas_natural",
+            new LukasChairBlock(AbstractBlock.Settings.copy(BAMBOO_PLANKS).nonOpaque()));
+    public static final Block BIRCH_CHAIR_LUKAS_NATURAL = registerBlock("birch_chair_lukas_natural",
+            new LukasChairBlock(AbstractBlock.Settings.copy(BIRCH_PLANKS).nonOpaque()));
+    public static final Block CHERRY_CHAIR_LUKAS_NATURAL = registerBlock("cherry_chair_lukas_natural",
+            new LukasChairBlock(AbstractBlock.Settings.copy(CHERRY_PLANKS).nonOpaque()));
+    public static final Block CRIMSON_CHAIR_LUKAS_NATURAL = registerBlock("crimson_chair_lukas_natural",
+            new LukasChairBlock(AbstractBlock.Settings.copy(CRIMSON_PLANKS).nonOpaque()));
+    public static final Block DARK_OAK_CHAIR_LUKAS_NATURAL = registerBlock("dark_oak_chair_lukas_natural",
+            new LukasChairBlock(AbstractBlock.Settings.copy(DARK_OAK_PLANKS).nonOpaque()));
+    public static final Block JUNGLE_CHAIR_LUKAS_NATURAL = registerBlock("jungle_chair_lukas_natural",
+            new LukasChairBlock(AbstractBlock.Settings.copy(JUNGLE_PLANKS).nonOpaque()));
+    public static final Block MANGROVE_CHAIR_LUKAS_NATURAL = registerBlock("mangrove_chair_lukas_natural",
+            new LukasChairBlock(AbstractBlock.Settings.copy(MANGROVE_PLANKS).nonOpaque()));
+    public static final Block OAK_CHAIR_LUKAS_NATURAL = registerBlock("oak_chair_lukas_natural",
+            new LukasChairBlock(AbstractBlock.Settings.copy(OAK_PLANKS).nonOpaque()));
+    public static final Block SPRUCE_CHAIR_LUKAS_NATURAL = registerBlock("spruce_chair_lukas_natural",
+            new LukasChairBlock(AbstractBlock.Settings.copy(SPRUCE_PLANKS).nonOpaque()));
+    public static final Block WARPED_CHAIR_LUKAS_NATURAL = registerBlock("warped_chair_lukas_natural",
+            new LukasChairBlock(AbstractBlock.Settings.copy(WARPED_PLANKS).nonOpaque()));
     public static final Block ACACIA_LOG_SEAT = registerBlock("acacia_log_seat",
             new LogSeat(AbstractBlock.Settings.copy(ACACIA_PLANKS).nonOpaque()));
     public static final Block BIRCH_LOG_SEAT = registerBlock("birch_log_seat",