+++ src/main/java/com/nemonotfound/nemoscarpentry/datagen/RecipeGenerator.java
@@ -201,6 +201,8 @@ public class RecipeGenerator extends FabricRecipeProvider {
 
         createCarpentryRecipe(exporter, List.of(Pair.of(Blocks.CHERRY_PLANKS, 2)),
                 ModItems.CHERRY_GLASS_DOOR_FRAME, 2, CarpentryTools.SAW.asString());
+
+        createChairLukasNaturalRecipes(exporter);
     }
 
     private void createTableTheClassicRecipes(RecipeExporter exporter) {
@@ -226,6 +228,31 @@ public class RecipeGenerator extends FabricRecipeProvider {
                 ModBlocks.SPRUCE_TABLE_THE_CLASSIC, 1, CarpentryTools.SAW.asString());
         createCarpentryRecipe(exporter, List.of(Pair.of(Blocks.WARPED_PLANKS, 1)),
                 ModBlocks.WARPED_TABLE_THE_CLASSIC, 1, CarpentryTools.SAW.asString());
+    }
+
+    private void createChairLukasNaturalRecipes(RecipeExporter exporter) {
+        createCarpentryRecipe(exporter, Pair.of(Blocks.ACACIA_PLANKS, 1), Pair.of(ItemTags.ACACIA_LOGS, 1),
+                ModItems.ACACIA_CHAIR_LUKAS_NATURAL, 1, CarpentryTools.SAW.asString());
+        createCarpentryRecipe(exporter, Pair.of(Blocks.BAMBOO_PLANKS, 1), Pair.of(ItemTags.BAMBOO_BLOCKS, 1),
+                ModItems.BAMBOO_CHAIR_LUKAS_NATURAL, 1, CarpentryTools.SAW.asString());
+        createCarpentryRecipe(exporter, Pair.of(Blocks.BIRCH_PLANKS, 1), Pair.of(ItemTags.BIRCH_LOGS, 1),
+                ModItems.BIRCH_CHAIR_LUKAS_NATURAL, 1, CarpentryTools.SAW.asString());
+        createCarpentryRecipe(exporter, Pair.of(Blocks.CHERRY_PLANKS, 1), Pair.of(ItemTags.CHERRY_LOGS, 1),
+                ModItems.CHERRY_CHAIR_LUKAS_NATURAL, 1, CarpentryTools.SAW.asString());
+        createCarpentryRecipe(exporter, Pair.of(Blocks.CRIMSON_PLANKS, 1), Pair.of(ItemTags.CRIMSON_STEMS, 1),
+                ModItems.CRIMSON_CHAIR_LUKAS_NATURAL, 1, CarpentryTools.SAW.asString());
+        createCarpentryRecipe(exporter, Pair.of(Blocks.DARK_OAK_PLANKS, 1), Pair.of(ItemTags.DARK_OAK_LOGS, 1),
+                ModItems.DARK_OAK_CHAIR_LUKAS_NATURAL, 1, CarpentryTools.SAW.asString());
+        createCarpentryRecipe(exporter, Pair.of(Blocks.JUNGLE_PLANKS, 1), Pair.of(ItemTags.JUNGLE_LOGS, 1),
+                ModItems.JUNGLE_CHAIR_LUKAS_NATURAL, 1, CarpentryTools.SAW.asString());
+        createCarpentryRecipe(exporter, Pair.of(Blocks.MANGROVE_PLANKS, 1), Pair.of(ItemTags.MANGROVE_LOGS, 1),
+                ModItems.MANGROVE_CHAIR_LUKAS_NATURAL, 1, CarpentryTools.SAW.asString());
+        createCarpentryRecipe(exporter, Pair.of(Blocks.OAK_PLANKS, 1), Pair.of(ItemTags.OAK_LOGS, 1),
+                ModItems.OAK_CHAIR_LUKAS_NATURAL, 1, CarpentryTools.SAW.asString());
+        createCarpentryRecipe(exporter, Pair.of(Blocks.SPRUCE_PLANKS, 1), Pair.of(ItemTags.SPRUCE_LOGS, 1),
+                ModItems.SPRUCE_CHAIR_LUKAS_NATURAL, 1, CarpentryTools.SAW.asString());
+        createCarpentryRecipe(exporter, Pair.of(Blocks.WARPED_PLANKS, 1), Pair.of(ItemTags.WARPED_STEMS, 1),
+                ModItems.WARPED_CHAIR_LUKAS_NATURAL, 1, CarpentryTools.SAW.asString());
     }
 
     private void createSawRecipe(RecipeExporter exporter) {
@@ -269,6 +296,23 @@ public class RecipeGenerator extends FabricRecipeProvider {
                 .create(RecipeCategory.DECORATIONS, output, outputCount, tool);
 
         addIngredientsToRecipe(builder, ingredients);
+
+        builder.criterion(FabricRecipeProvider.hasItem(output), conditionsFromItem(output))
+                .offerTo(exporter, new Identifier(getRecipeName(output) + "_carpentry"));
+    }
+
+    private void createCarpentryRecipe(RecipeExporter exporter, Pair<Block, Integer> ingredientPair,
+                                       Pair<TagKey<Item>, Integer> secondIngredientPair, ItemConvertible output,
+                                       int outputCount, String tool) {
+        Block firstIngredient = ingredientPair.getFirst();
+        TagKey<Item> secondIngredient = secondIngredientPair.getFirst();
+
+        CarpentryRecipeJsonBuilder builder = CarpentryRecipeJsonBuilder
+                .create(RecipeCategory.DECORATIONS, output, outputCount, tool)
+                .input(Ingredient.ofItems(firstIngredient), ingredientPair.getSecond())
+                .criterion(hasItem(firstIngredient), conditionsFromItem(firstIngredient))
+                .input(Ingredient.fromTag(secondIngredientPair.getFirst()), secondIngredientPair.getSecond())
+                .criterion("has_logs", conditionsFromTag(secondIngredient));
 
         builder.criterion(FabricRecipeProvider.hasItem(output), conditionsFromItem(output))
                 .offerTo(exporter, new Identifier(getRecipeName(output) + "_carpentry"));